requirements.txt
streamlit==1.32.0
yfinance==0.2.37
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
xgboost==2.0.3
scikit-learn==1.4.1.post1
google-generativeai==0.4.1